feat: 对 P-Agent 安装依赖更新场景提供文件原子操作支持 (closed #2194)
