+++ apps/node_man/models.py
@@ -132,8 +132,6 @@ class GlobalSettings(models.Model):
         INSTALL_OTHER_AGENT_V1_BIZ_BLACKLIST = "INSTALL_OTHER_AGENT_V1_BIZ_BLACKLIST"
         INSTALL_OTHER_AGENT_V2_BIZ_BLACKLIST = "INSTALL_OTHER_AGENT_V2_BIZ_BLACKLIST"
         NEED_TO_WAIT_EXTRA_INSTALL_COMPLETE = "NEED_TO_WAIT_EXTRA_INSTALL_COMPLETE"
-        # P-Agent 安装脚本名称
-        SETUP_PAGENT_SCRIPT_FILENAME = "SETUP_PAGENT_SCRIPT_FILENAME"
         # 周期删除订阅任务数据
         CLEAN_SUBSCRIPTION_DATA_MAP = "CLEAN_SUBSCRIPTION_DATA_MAP"
         # 是否开启Agent包管理
